Move tabs.js and helpers.js into platform/common
