+++ src/platform/common/helpers.js
