+++ src/platform/common/tabs.js
